Move the boilerplate init scenario in a single place
