+++ test/oauth/password.test.js
@@ -3,62 +3,14 @@ const should = require('should');
 
 const app = require('./bootstrap');
 const services = require('../../lib/services');
-const db = require('../../lib/db');
-const checkTokenResponse = require('./checkTokenResponse');
+const { checkTokenResponse, createOAuthScenario } = require('./testUtils');
 
-const credentialService = services.credential;
-const userService = services.user;
-const applicationService = services.application;
 const tokenService = services.token;
 
 describe('Functional Test Client Password grant', function () {
-  let fromDbUser1, fromDbApp, refreshToken;
+  let fromDbApp, refreshToken;
 
-  const user1 = {
-    username: 'irfanbaqui',
-    firstname: 'irfan',
-    lastname: 'baqui',
-    email: 'irfan@eg.com'
-  };
-
-  const user2 = {
-    username: 'somejoe',
-    firstname: 'joe',
-    lastname: 'smith',
-    email: 'joe@eg.com'
-  };
-
-  before(() =>
-    db.flushdb()
-      .then(() => Promise.all([userService.insert(user1), userService.insert(user2)]))
-      .then(([_fromDbUser1, _fromDbUser2]) => {
-        should.exist(_fromDbUser1);
-        should.exist(_fromDbUser2);
-
-        fromDbUser1 = _fromDbUser1;
-
-        const app1 = {
-          name: 'irfan_app',
-          redirectUri: 'https://some.host.com/some/route'
-        };
-
-        return applicationService.insert(app1, fromDbUser1.id);
-      })
-      .then(_fromDbApp => {
-        should.exist(_fromDbApp);
-        fromDbApp = _fromDbApp;
-
-        return credentialService.insertScopes(['someScope']);
-      })
-      .then(() => Promise.all([
-        credentialService.insertCredential(fromDbUser1.id, 'basic-auth', { password: 'user-secret' }),
-        credentialService.insertCredential(fromDbApp.id, 'oauth2', { secret: 'app-secret', scopes: ['someScope'] })
-      ]))
-      .then(([userRes, appRes]) => {
-        should.exist(userRes);
-        should.exist(appRes);
-      })
-  );
+  before(() => createOAuthScenario().then(([user, app]) => { fromDbApp = app; }));
 
   it('should grant access token when no scopes are specified', function (done) {
     const request = session(app);
